feat: decouple buffer size of pitch detection from sound input
the frame count which fires every update of the sound input (PAWrapper) should be as low as possible. The main reason for this is that there will be features that will be built on top of these updates which work better with continuous streams of data (i.e. graphic updates based on sound input).